internal: rename tiltfile2 pkg -> tiltfile
