+++ internal/demo/script.go
@@ -18,7 +18,7 @@ import (
 	"github.com/windmilleng/tilt/internal/logger"
 	"github.com/windmilleng/tilt/internal/model"
 	"github.com/windmilleng/tilt/internal/store"
-	"github.com/windmilleng/tilt/internal/tiltfile2"
+	"github.com/windmilleng/tilt/internal/tiltfile"
 	"golang.org/x/sync/errgroup"
 	"k8s.io/api/core/v1"
 )
@@ -167,8 +167,8 @@ func (s Script) Run(ctx context.Context) error {
 			return ctx.Err()
 		}
 
-		tfPath := filepath.Join(dir, tiltfile2.FileName)
-		manifests, _, _, err := tiltfile2.Load(ctx, tfPath, nil)
+		tfPath := filepath.Join(dir, tiltfile.FileName)
+		manifests, _, _, err := tiltfile.Load(ctx, tfPath, nil)
 		if err != nil {
 			return err
 		}
