+++ internal/engine/configs_controller.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/windmilleng/tilt/internal/store"
-	"github.com/windmilleng/tilt/internal/tiltfile2"
+	"github.com/windmilleng/tilt/internal/tiltfile"
 )
 
 type ConfigsController struct {
@@ -45,7 +45,7 @@ func (cc *ConfigsController) OnChange(ctx context.Context, st store.RStore) {
 		for _, m := range initManifests {
 			matching[string(m)] = true
 		}
-		manifests, globalYAML, configFiles, err := tiltfile2.Load(ctx, tiltfile2.FileName, matching)
+		manifests, globalYAML, configFiles, err := tiltfile.Load(ctx, tiltfile.FileName, matching)
 		st.Dispatch(ConfigsReloadedAction{
 			Manifests:   manifests,
 			GlobalYAML:  globalYAML,
