+++ internal/tiltfile/docker_compose.go
@@ -1,4 +1,4 @@
-package tiltfile2
+package tiltfile
 
 import (
 	"fmt"
