+++ internal/tiltfile/k8s.go
@@ -1,4 +1,4 @@
-package tiltfile2
+package tiltfile
 
 import (
 	"fmt"
