+++ internal/tiltfile/tiltfile_docker_compose_test.go
@@ -1,4 +1,4 @@
-package tiltfile2
+package tiltfile
 
 import (
 	"fmt"
